requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.2
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyPDF2==3.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
